+++ psi-mod-CV-browser/src/components/smiles/SmilesView.vue
