+++ psi-mod-CV-browser/src/pages/index.vue
@@ -3,5 +3,5 @@
 </template>
 
 <script lang="ts" setup>
-import ProteinModificationsTable from '@/components/ProteinModificationsTable.vue'
+import ProteinModificationsTable from '@/components/table/ProteinModificationsTable.vue'
 </script>
